Breaking changes for 4.0.0 release (#20)
This commit removes `init`, `update` and `subscriptions` function from the `Glue` data type.
This also means that we no longer need `a` type argument.

Reason for this change is that in real-world application one has to often pass an extra argument to these functions. The original intention was to extract those arguments from the parent model using `glue` constructor however this seems to not be obvious to people and in some situations, it might be even impossible. Usually this end ups by folks defining a function to construct the `Glue` type like
`Flags -> Glue....` and constructing the type over and over whenever it's needed.

This breaking change discourages such pattern. The glue should be a constant value unless the number of submodules in the parent module varies based on some state.